refactoring to change name from host_switcher to api_modifler
