+++ requestly-android-core/src/main/java/io/requestly/android/core/modules/apiModifier/models/ActionProcessor.kt
@@ -1,4 +1,4 @@
-package io.requestly.android.core.modules.hostSwitcher.models
+package io.requestly.android.core.modules.apiModifier.models
 
 import okhttp3.Request
